requirement.txt
torch==1.3.0
torchvision==0.4.2
skimage==0.15.0
opencv==4.1.2
pydensecrf==1.0
matplotlib==3.1.1
pandas==0.25.2
pillow==6.2.1
numpy==1.17.2
